# api.yaml
openapi: 3.0.0
info:
  title: API управления устройствами
  version: 1.0.0

components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
    Device:
      type: object
      properties:
        device_id:
          type: string
        deviceType:
          type: string
        serialNumber:
          type: string
        status:
          type: string

paths:
  /devices/{device_id}:
    get:
      summary: Получение информации об устройстве
      description: Возвращает подробную информацию о конкретном устройстве по его ID.
      parameters:
        - name: device_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'
        '404':
          description: Устройство не найдено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /devices/{device_id}/status:
    put:
      summary: Обновление состояния устройства
      description: Позволяет изменить состояние устройства (например, включить/выключить).
      parameters:
        - name: device_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
        '404':
          description: Устройство не найдено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /devices/{device_id}/commands:
    post:
      summary: Отправка команды устройству
      description: Отправляет команду устройству (например "открыть ворота").
      parameters:
        - name: device_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                command:
                  type: string
                value:
                  type: string
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                type: object
                properties:
                  result:
                    type: string
        '404':
          description: Устройство не найдено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /devices:
    post:
      summary: Добавление устройства
      description: Добавляет новое устройство.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                deviceType:
                  type: string
                serialNumber:
                  type: string
                status:
                  type: string
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'